refactor(FAQ): Extract common code to parent
