+++ src/app/help/faq/faq.component.scss
@@ -1,4 +1,3 @@
-
 .title-icon {
   margin-right: 8px;
 }
