+++ src/app/help/faq/student-faq/student-faq.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+import { FaqComponent } from '../faq.component';
+
+@Component({
+  selector: 'app-student-faq',
+  templateUrl: './student-faq.component.html',
+  styleUrls: ['../faq.component.scss']
+})
+export class StudentFaqComponent extends FaqComponent {}
