+++ src/app/help/faq/teacher-faq/teacher-faq.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+import { FaqComponent } from '../faq.component';
+
+@Component({
+  selector: 'app-teacher-faq',
+  templateUrl: './teacher-faq.component.html',
+  styleUrls: ['../faq.component.scss']
+})
+export class TeacherFaqComponent extends FaqComponent {}
