+++ src/app/help/help-routing.module.ts
@@ -3,8 +3,8 @@ import { RouterModule, Routes } from '@angular/router';
 import { HelpComponent } from './help.component';
 import { HelpHomeComponent } from './help-home/help-home.component';
 import { GettingStartedComponent } from './getting-started/getting-started.component';
-import { TeacherFaqComponent } from './teacher-faq/teacher-faq.component';
-import { StudentFaqComponent } from './student-faq/student-faq.component';
+import { TeacherFaqComponent } from './faq/teacher-faq/teacher-faq.component';
+import { StudentFaqComponent } from './faq/student-faq/student-faq.component';
 
 const helpRoutes: Routes = [
   {
