--- src/app/help/teacher-faq/teacher-faq.component.scss
@@ -1,16 +1,0 @@
-
-.title-icon {
-  margin-right: 8px;
-}
-
-.link {
-  cursor: pointer;
-}
-
-.indented-text {
-  margin-left: 24px;
-}
-
-.warning {
-  color: red;
-}
